Add Actor Model
We have successfully added a way for us to process messages in
order by consuming them from some sort of Inbox. We have also
created a dummy Inbox that uses some Interval to prove that we
can do things async. We can now focus on creating Plugins such
as an RMQ Inbox or perhaps a ThreadInbox that spins up some
child Actors as threads.